# backend/public/openapi.yml
openapi: 3.0.0
info:
  title: Loyalty Haven API
  description: API for managing room bookings, payments, and user authentication.
  version: 1.0.0
servers:
  - url: http://localhost:5000
    description: Local development server
security:
  - cookieAuth: []
...
paths:
...
  /api/admin/register:
    post:
      tags:
        - Admin
      summary: Register a new admin
      description: |
        Register a new admin. Requires an admin token from another admin.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminRegisterRequest"
      responses:
        "200":
          description: Admin registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminRegisterResponse"
  /api/admin/get-token:
    get:
      tags:
        - Admin
      summary: Get an admin token
      description: |
        Retrieve a token to register a new admin. Requires an admin account.
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Admin token retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
  /api/admin/all-rooms:
    get:
      tags:
        - Admin
      summary: Get all rooms
      description: |
        Retrieve all rooms in the system. Requires an admin account.
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: sort
          in: query
          schema:
            type: string
      security:
        - cookieAuth: []
      responses:
        "200":
          description: List of rooms
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AllRoomsResponse"
  /api/admin/all-bookings:
    get:
      tags:
        - Admin
      summary: Get all bookings
      description: |
        Retrieve all bookings in the system. Requires an admin account.
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: sort
          in: query
          schema:
            type: string
      security:
        - cookieAuth: []
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AllBookingsResponse"
  /api/admin/all-users:
    get:
      tags:
        - Admin
      summary: Get all users
      description: |
        Retrieve all users in the system. Requires an admin account.
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: sort
          in: query
          schema:
            type: string
      security:
        - cookieAuth: []
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AllUsersResponse"
  /api/admin/activate-room:
    put:
      tags:
        - Admin
      summary: Activate a room
      description: |
        Activate a room so it can be booked. Requires an admin account.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivateRoomRequest"
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Room activated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Room activated successfully"
  /api/admin/deactivate-room:
    put:
      tags:
        - Admin
      summary: Deactivate a room
      description: |
        Deactivate a room so it cannot be booked. Requires an admin account.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeactivateRoomRequest"
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Room deactivated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Room deactivated successfully"
  /api/admin/revoke-owner:
    put:
      tags:
        - Admin
      summary: Revoke owner access
      description: |
        Revoke a user's room owner access. Requires an admin account.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RevokeOwnerRequest"
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Owner access revoked successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Room owner access revoked successfully"
  /api/admin/grant-owner:
    put:
      tags:
        - Admin
      summary: Grant owner access
      description: |
        Grant a user room owner access. Requires an admin account.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GrantOwnerRequest"
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Owner access granted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Room owner access granted successfully"
...
components:
  schemas:
...
    AllRoomsResponse:
      type: object
      properties:
        meta:
          type: object
          properties:
            sort:
              type: string
            totalPages:
              type: number
            pageSize:
              type: number
            page:
              type: number
        data:
          type: object
          properties:
            rooms:
              type: array
              items:
                $ref: "#/components/schemas/Room"
...
    AdminRegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        name:
          type: string
        phone:
          type: string
        PhoneOTP:
          type: string
        EmailOTP:
          type: string
        admin_token:
          type: string
      required:
        - email
        - password
        - name
        - phone
        - PhoneOTP
        - EmailOTP
        - admin_token
    AdminRegisterResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        accountType:
          type: string
        message:
          type: string
          example: "Logged in successfully."
    AllBookingsResponse:
      type: object
      properties:
        meta:
          type: object
          properties:
            sort:
              type: string
            totalPages:
              type: number
            pageSize:
              type: number
            page:
              type: number
        data:
          type: object
          properties:
            bookings:
              type: array
              items:
                $ref: "#/components/schemas/Booking"
    AllUsersResponse:
      type: object
      properties:
        meta:
          type: object
          properties:
            sort:
              type: string
            totalPages:
              type: number
            pageSize:
              type: number
            page:
              type: number
        data:
          type: object
          properties:
            users:
              type: array
              items:
                $ref: "#/components/schemas/User"
    ActivateRoomRequest:
      type: object
      properties:
        roomId:
          type: string
      required:
        - roomId
    DeactivateRoomRequest:
      type: object
      properties:
        roomId:
          type: string
      required:
        - roomId
    RevokeOwnerRequest:
      type: object
      properties:
        userId:
          type: string
      required:
        - userId
    GrantOwnerRequest:
      type: object
      properties:
        userId:
          type: string
      required:
        - userId
...
    User:
      $ref: "#/components/schemas/UserSchema"
    Room:
      $ref: "#/components/schemas/RoomSchema"
    Booking:
      $ref: "#/components/schemas/BookingSchema"
...
    UserSchema:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        accountType:
          type: string
          enum: [booker, owner, admin]
        accountStatus:
          type: string
          enum: [active, banned, dormant]
        bookings:
          type: array
          items:
            type: string
        rooms:
          type: array
          items:
            type: string
        phone:
          type: string
      required:
        - name
        - email
        - password
        - accountType
        - accountStatus
        - phone
    RoomSchema:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        area:
          type: number
        capacity:
          type: number
        price_per_hour:
          type: number
        street:
          type: string
        location:
          type: string
        availability:
          type: string
        amenities:
          type: array
          items:
            type: string
        owner:
          type: string
        total_hours_booked:
          type: number
        images:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, inactive, deleted]
      required:
        - name
        - price_per_hour
        - location
        - owner
    BookingSchema:
      type: object
      properties:
        bookerId:
          type: string
        roomId:
          type: string
        status:
          type: string
          enum: [pending, cancelled, completed, active]
        checkIn:
          type: string
          format: date-time
        checkOut:
          type: string
          format: date-time
        duration:
          type: number
        fees:
          type: number
      required:
        - bookerId
        - roomId
        - checkIn
        - checkOut
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: lh_token